backend/requirements.txt
fastapi
uvicorn
python-multipart
chromadb
google-generativeai
groq
python-dotenv
pypdf
